tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/006.html

<!--
var x;
var y;
var z;
try {
  importScripts('data:text/javascript,x=1',
                'data:text/javascript,throw 2',
                'data:text/javascript,z=3');
} catch(e) {
  y = e;
}
postMessage([x, y, z]);
/*
-->
<!doctype html>
<title>importScripts uncaught exception</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data[0], 1);
    assert_equals(e.data[1], 2);
    assert_equals(e.data[2], undefined);
    this.done();
  });
});
</script>
<!--
*/
//-->
