LayoutTests/imported/w3c/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/006-expected.txt

CONSOLE MESSAGE: line 12: DataCloneError: The object can not be cloned.

Harness Error (FAIL), message = DataCloneError: The object can not be cloned.

TIMEOUT importScripts uncaught exception Test timed out
